testing python error

+++ checkHumidity.py
@@ -23,7 +23,7 @@ if soil[0] == True:
             print data[1]
             print data[5]
             a = int(data[5])
-            if a < soil[1] and int(data[1]) < 2:
+            if a < soil[1] and float(data[1]) < 2:
                 print 'watering'
                 execfile('pumpStart.py')
                 print a;
